refactor(commons): align object name with file name
